Revert "HUE-7738 [editor] Extract shared type logic from the parsers into one place"
This reverts commit d5da9773e9bd4b1105e4467b23acb21007502222.
